forward the target of a redirection to the grabber
The website 01net.com uses feedsportal.com to deliver there feeds. The
article URL within the feeds points to feedsportal.com and is HTTP 301
forwarded to 01net.com. As the content grabber does not get notified
about this forwarding and the expected rule file is still
feedsportal.com.php. That isn't something the user expects and it could
be impossible to write content grabber rules for different websites
using feedsportal.com.

This was reported in fguillot/miniflux#267