atom: return only site urls without or empty rel attribute
The old behavior was to return any link if we ask for links with an
empty rel.

As far as I can see, the fallback functionality in getUrl is only used
by findSiteUrl and is needed to be compliant with RFC4287:

atom:link elements MAY have a "rel" attribute that indicates the link
elation type. If the "rel" attribute is not present, the link
element MUST be interpreted as if the link relation type is "alternate".