curl: disable SSLv3 only for known good libcurl versions
Disable SSLv3 by enforcing TLSv1.x for curl >= 7.34.0 and < 7.39.0.

Versions prior to 7.34 and at least when compiled against openssl
interpret this parameter as "limit to TLSv1.0" which fails for sites
which enforce TLS 1.1+.

Starting with curl 7.39.0 SSLv3 is disabled by default.

Set the user-agent using the curl options instead of creating a raw
header.