RSS 2.0: Use the managingEditor or dc:creator as the feed author
According to the specs, the managingEditor is responsible for editorial
content and the webMaster is responsible for technical issues.

Since the RSS 2.0 specs says the managingEditor element should contain an email
address, add support for the dc:creator which doesn't have this limitation.

Fixes feed author discovery of fixture geekstammtisch.de_episodes.mp3.rss.