RSS 2.0: Drop support for RSS 1.0 specific elements
Drop support for the dc:date element from the Dublin Core namespace. The
published element is the matching native RSS element. I assume the dc:date was
meant for use with RSS 1.0, since there is no similar "native" element.

Keep the dc:creator element, since the RSS 2.0 specs the author element should
contain the email address of the author of the item and not an author name.

Drop support for the updated element of the atom namespace. This element isn't
as elementary as the atom link element and finally, if someone needs atom
elements => use the atom format.