use XPath with propper namespace support for feed parsing (#112)
Beside a few exceptions, the RSS 1.0 class is a copy of the RSS 2.0 class,
but with support for the RSS 1.0 default Namespace.

XPath switch:
The returned values of xPath queries are more predictable. It's either
an empty array (no matches) or an array of SimpleXMLElements (matches
found).

The SimpleXMLElement object access notation returns in opposite an empty
object for elements which doesn't exist or NULL if the the parent node
of a requested element doesn't exist.

Better compatibility:
  - Atom/RSS nodes which don't reside in the default namespace are found

Possible incompatibilities:
It's very likely that namespaced nodes aren't found (any long) if:

- a prefix has no namespace uri
- feed uses a wrong namespace uri for any namespace